scripts/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
deprecation==2.1.0
frozenlist==1.5.0
gotrue==2.11.0
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
msgpack==1.1.0
multidict==6.1.0
packaging==24.2
postgrest==0.18.0
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
pynvim==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.0.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
yarl==1.18.3
